Rename the QueryBoundaryItemWrapper file to align with the actual name
Must've forgotten to do this in the components v7 migration.
